dockerisation/demos/ml-system/api/requirements.txt
flask==2.3.0
flask-cors==4.0.0
pandas==2.0.0
scikit-learn==1.3.0
psycopg2-binary==2.9.6
joblib==1.3.0